add more godoc strings
